rutas y vistas protegidas

+++ resources/views/admin/categories/index.blade.php
@@ -15,7 +15,9 @@
                                     <th width="20px">ID</th>
                                     <th >Nombre</th>
                                     <th colspan="3" class="text-center">
-                                        <a href="{{ route('categories.create') }}" class="btn btn-sm btn-link text-decoration-none text-primary">Añadir+</a>
+                                        @can('categories.create')
+                                            <a href="{{ route('categories.create') }}" class="btn btn-sm btn-link text-decoration-none text-primary">Añadir+</a>
+                                        @endcan
                                     </th>
                                 </tr>
                             </thead>
@@ -24,22 +26,28 @@
                                     <tr>
                                         <td scope="row">{{ $category->id }}</td>
                                         <td>{{ $category->name }}</td>
-                                        <td width='30px'class="px-1">
-                                            <a href="{{ route('categories.show', $category->id) }}" class="btn btn-sm btn-link text-decoration-none text-dark">Ver</a>
-                                        </td>
-                                        <td width='30px'class="px-1">
-                                            <a href="{{ route('categories.edit', $category->id) }}" class="btn btn-sm btn-link text-decoration-none text-secondary">Editar</a>
-                                        </td>
-                                        <td width='30px'class="px-1">
-                                            {!! Form::open(['route' => ['categories.destroy', $category->id], 'method' => 'DELETE']) !!}
-                                                <button class="btn btn-sm  btn-link text-decoration-none text-danger">Eliminar</button>
-                                            {!! Form::close() !!}
-                                        </td>
+                                        @can('categories.show')
+                                            <td width='30px'class="px-1">
+                                                <a href="{{ route('categories.show', $category->id) }}" class="btn btn-sm btn-link text-decoration-none text-dark">Ver</a>
+                                            </td>
+                                        @endcan
+                                        @can('categories.edit')
+                                            <td width='30px'class="px-1">
+                                                <a href="{{ route('categories.edit', $category->id) }}" class="btn btn-sm btn-link text-decoration-none text-secondary">Editar</a>
+                                            </td>
+                                        @endcan
+                                        @can('categories.destroy')
+                                            <td width='30px'class="px-1">
+                                                {!! Form::open(['route' => ['categories.destroy', $category->id], 'method' => 'DELETE']) !!}
+                                                    <button class="btn btn-sm  btn-link text-decoration-none text-danger">Eliminar</button>
+                                                {!! Form::close() !!}
+                                            </td>
+                                        @endcan
                                     </tr>
                                 @endforeach
                             </tbody>
                         </table>
-
+                            
                         {{ $categories->links() }}
                     </div>
                 </div>
